Add ability to create builds and stream build output
This is mostly following the example of `Releases`, but with a few build-specific tweaks. Once you create a build it's useful to stream the output so you can see when it finishes:

```
build = app.create_build('https://api.github.com/repos/cityofaustin/janis/tarball/master')
for line in build.stream_output():
  print(line)
```

It didn't seem like there was an easy way to add tests for these functions without adding a bunch of plumbing so I skipped that part. Let me know if there's something I should add for tests.